Improve filter factory
